requirements.txt
python-dotenv~=1.1.1
python-telegram-bot~=22.3
openai~=1.97.1
aiohttp~=3.12.14